Refactor: More cleanups of duplicates
- Remove some more duplicate tests via another Mixin
- Remove some near duplicate by additional DDT testing
